test(streaming): fold the short-dictation tests into one file
Two files, two near-identical 55-line fake sockets, and eight tests for two
fixes. Now one harness and four tests, each pinning something that can
actually regress:

- Freestyle Cloud holds audio captured while its socket opens (was the
  provider most affected, and the only one of the three with no direct
  coverage before).
- OpenAI holds until the session is *configured*, not merely open — the one
  provider whose gate differs.
- Soniox waits for tokens rather than resolving empty the moment it
  finalizes.
- Silence still resolves promptly, so that guard can't turn into an 8s hang.

Dropped: Deepgram's buffering test (same shared holder as the others, same
shape), a separate ordering test (the kept one already asserts exact order),
and a cancel-clears test.
